1. Factored out genesis hawk job.
