+++ iris-mpc-cpu/src/genesis/mod.rs
@@ -1,6 +1,8 @@
 mod batch_generator;
 mod hawk_handle;
+mod hawk_job;
 mod utils;
 
 pub use batch_generator::{BatchGenerator, BatchIterator};
-pub use hawk_handle::{Handle, Job, JobRequest, JobResult};
+pub use hawk_handle::Handle;
+pub use hawk_job::{Job, JobRequest, JobResult};
